+++ cm/cmind/__init__.py
@@ -2,7 +2,7 @@
 #
 # Written by Grigori Fursin
 
-__version__ = "2.3.8"
+__version__ = "2.4.0"
 
 from cmind.core import access
 from cmind.core import error